# BUG: stats.qmc: Avoid infinite loop in sobol when requesting more than 2**32 samples