--- a/scipy/stats/_sobol.pyx
+++ b/scipy/stats/_sobol.pyx
@@ -444,7 +444,8 @@ def _test_find_index(p_cumulative, size, value):
     """Wrapper for testing in python"""
     return _find_index(p_cumulative, size, value)
 
-cpdef _test_low_0_bit(cnp.uint64_t x):
+
+def _test_low_0_bit(cnp.uint64_t x):
     # type: (int,) -> int
     """Wrapper for testing in python"""
     return low_0_bit(x)