--- a/scipy/stats/tests/test_qmc.py
+++ b/scipy/stats/tests/test_qmc.py
@@ -9,7 +9,7 @@ from numpy.testing import (assert_allclose, assert_equal, assert_array_equal,
 
 from scipy.spatial import distance
 from scipy.stats import shapiro
-from scipy.stats._sobol import _test_find_index
+from scipy.stats._sobol import _test_find_index, _test_low_0_bit
 from scipy.stats import qmc
 from scipy.stats._qmc import (
     van_der_corput, n_primes, primes_from_2_to,
@@ -881,6 +881,29 @@ class TestSobol(QMCEngineTests):
         engine = qmc.Sobol(2, scramble=False, bits=64)
         sample = engine.random(8)
         assert_array_equal(self.unscramble_nd, sample)
+
+
+
+class TestLow0Bit:
+    def test_examples(self):
+        test_vector = [
+            # from low_0_bit's docstring
+            (0b0000, 1),
+            (0b0001, 2),
+            (0b0010, 1),
+            (0b0101, 2),
+            (0b0111, 4),
+            # gh-23409
+            # (2 ** 32 - 1, 33),
+            # (2 ** 32, 1),
+            # (2 ** 33 - 1, 34),
+            # (2 ** 64 - 1, 65),
+        ]
+        for in_, out in test_vector:
+            # Try both specializations of this function if the argument fits
+            if in_ <= 2 ** 32 - 1:
+                assert_equal(_test_low_0_bit['uint32_t'](in_), out)
+            assert_equal(_test_low_0_bit['uint64_t'](in_), out)
 
 
 class TestPoisson(QMCEngineTests):