Fix bug where low_0_bit could loop forever. Also, add tests for low_0_bit.

#### Reference issue

Closes #23409 

#### What does this implement/fix?

I changed the following things:

* When evaluating a 64-bit Sobol sample, the bit shift was done in 32 bits, before the 32-bit result was promoted to a 64-bit number. The shift is now always done in the same type as `uint_32_64`.
* When evaluating `low_0_bit` where x was all ones, the program would loop forever, because `1 << x` can wrap around when x is larger than the number of bits in the type. `low_0_bit` now shifts x right one place each loop, so it is guaranteed to eventually reach zero.
   * This changes the return value for `low_0_bit(2**64 - 1)` to 65. I think this is the most natural return value in this situation. For all previous values of x, `low_0_bit(2**x - 1) == x + 1`.
* I experimented with a different change for fixing the infinite loop, which you can see in 6dbb7e5b43bd2a965b933dec001d5e66d2529da9. However, I felt it was too complicated, so I changed it to the version you see here.
* I marked `low_0_bit` as cpdef, so that I could call it from Python test code. I wrote some tests for it. I did this because testing this by running stats.qmc.Sobol with 2**32 samples would take too long.
* I verified that `_draw` calls the C function directly, rather than the Python wrapper.

#### Additional information

#### Performance notes

Since I was introducing a new loop-carried dependency, which can prevent optimization, I wanted to check that this bug fix didn't harm performance too much. I ran the BenchQMCSobol benchmark. In some circumstances, there is a 5% performance regression.

```
$ (cd benchmarks/; asv continuous --factor 1.05 --bench stats.BenchQMCSobol main 5df4c3653c21e30c0fde4c179132cbd24c49ab87 --interleave-rounds -a rounds=10)
[90.00%] · For scipy commit 5df4c365 <njo-sobol-loop> (round 10/10):
[90.00%] ·· Benchmarking virtualenv-py3.12-Cython-meson-python-numpy-pybind11-pytest-pythran
[95.00%] ··· stats.BenchQMCSobol.time_sobol                                                                          ok
[95.00%] ··· ===== ========= ========= ========= ============= ============ ============ =============
             --                                           base2                                       
             ----- -----------------------------------------------------------------------------------
               d       3         10        11          12           14           16            20     
             ===== ========= ========= ========= ============= ============ ============ =============
               1    224±2μs   237±3μs   244±3μs     256±4μs      339±3μs      662±9μs     12.1±0.07ms 
               50   262±3μs   364±6μs   460±3μs     666±7μs     6.79±0.8ms   17.3±0.5ms     222±1ms   
              100   301±4μs   485±6μs   688±6μs   4.83±0.09ms   9.61±0.7ms   30.5±0.4ms     432±3ms   
             ===== ========= ========= ========= ============= ============ ============ =============

[95.00%] · For scipy commit 3fe8b508 <main> (round 10/10):
[95.00%] ·· Building for virtualenv-py3.12-Cython-meson-python-numpy-pybind11-pytest-pythran...
[95.00%] ·· Benchmarking virtualenv-py3.12-Cython-meson-python-numpy-pybind11-pytest-pythran
[100.00%] ··· stats.BenchQMCSobol.time_sobol                                                                          ok
[100.00%] ··· ===== ======= =============
                d    base2               
              ----- ------- -------------
                1      3       227±6μs   
                1      10      239±5μs   
                1      11      248±3μs   
                1      12      260±5μs   
                1      14      336±3μs   
                1      16      630±6μs   
                1      20     11.7±0.1ms 
                50     3       265±4μs   
                50     10      363±5μs   
                50     11      466±6μs   
                50     12      669±9μs   
                50     14    6.90±0.07ms 
                50     16     17.5±0.1ms 
                50     20      224±3ms   
               100     3       317±5μs   
               100     10      499±5μs   
               100     11      689±10μs  
               100     12    4.82±0.08ms 
               100     14     9.76±0.8ms 
               100     16     30.6±0.3ms 
               100     20      438±5ms   
              ===== ======= =============

| Change   | Before [3fe8b508] <main>   | After [5df4c365] <njo-sobol-loop>   |   Ratio | Benchmark (Parameter)                  |
|----------|----------------------------|-------------------------------------|---------|----------------------------------------|
| +        | 630±6μs                    | 662±9μs                             |    1.05 | stats.BenchQMCSobol.time_sobol(1, 16)  |
| -        | 317±5μs                    | 301±4μs                             |    0.95 | stats.BenchQMCSobol.time_sobol(100, 3) |
```