@rgommers 2025-08-07
Then it may be good to move back to `cdef` and if you want to keep the test, which seems reasonable, add a separate utility `cpdef low_0_bit_py` which forwards to `low_0_bit`?